Re-organized src code: better naming, method placement
