+++ src/RcppExports.cpp
@@ -358,48 +358,6 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// add_coal_sites_cpp
-SEXP add_coal_sites_cpp(SEXP& ref_genome_ptr, const std::vector<arma::mat>& seg_sites, const arma::mat& Q, const std::vector<double>& pi_tcag, const std::vector<double>& insertion_rates, const std::vector<double>& deletion_rates, uint32 n_threads, const bool& show_progress);
-RcppExport SEXP _jackalope_add_coal_sites_cpp(SEXP ref_genome_ptrSEXP, SEXP seg_sitesSEXP, SEXP QSEXP, SEXP pi_tcagSEXP, SEXP insertion_ratesSEXP, SEXP deletion_ratesSEXP, SEXP n_threadsSEXP, SEXP show_progressSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< SEXP& >::type ref_genome_ptr(ref_genome_ptrSEXP);
-    Rcpp::traits::input_parameter< const std::vector<arma::mat>& >::type seg_sites(seg_sitesSEXP);
-    Rcpp::traits::input_parameter< const arma::mat& >::type Q(QSEXP);
-    Rcpp::traits::input_parameter< const std::vector<double>& >::type pi_tcag(pi_tcagSEXP);
-    Rcpp::traits::input_parameter< const std::vector<double>& >::type insertion_rates(insertion_ratesSEXP);
-    Rcpp::traits::input_parameter< const std::vector<double>& >::type deletion_rates(deletion_ratesSEXP);
-    Rcpp::traits::input_parameter< uint32 >::type n_threads(n_threadsSEXP);
-    Rcpp::traits::input_parameter< const bool& >::type show_progress(show_progressSEXP);
-    rcpp_result_gen = Rcpp::wrap(add_coal_sites_cpp(ref_genome_ptr, seg_sites, Q, pi_tcag, insertion_rates, deletion_rates, n_threads, show_progress));
-    return rcpp_result_gen;
-END_RCPP
-}
-// make_gamma_mats
-arma::field<arma::mat> make_gamma_mats(const std::vector<uint32>& seq_sizes, const uint32& gamma_size_, const double& shape);
-RcppExport SEXP _jackalope_make_gamma_mats(SEXP seq_sizesSEXP, SEXP gamma_size_SEXP, SEXP shapeSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const std::vector<uint32>& >::type seq_sizes(seq_sizesSEXP);
-    Rcpp::traits::input_parameter< const uint32& >::type gamma_size_(gamma_size_SEXP);
-    Rcpp::traits::input_parameter< const double& >::type shape(shapeSEXP);
-    rcpp_result_gen = Rcpp::wrap(make_gamma_mats(seq_sizes, gamma_size_, shape));
-    return rcpp_result_gen;
-END_RCPP
-}
-// check_gamma_mats
-void check_gamma_mats(const std::vector<arma::mat>& mats, const std::vector<uint32>& seq_sizes);
-RcppExport SEXP _jackalope_check_gamma_mats(SEXP matsSEXP, SEXP seq_sizesSEXP) {
-BEGIN_RCPP
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const std::vector<arma::mat>& >::type mats(matsSEXP);
-    Rcpp::traits::input_parameter< const std::vector<uint32>& >::type seq_sizes(seq_sizesSEXP);
-    check_gamma_mats(mats, seq_sizes);
-    return R_NilValue;
-END_RCPP
-}
 // phylo_info_to_trees
 SEXP phylo_info_to_trees(const List& genome_phylo_info);
 RcppExport SEXP _jackalope_phylo_info_to_trees(SEXP genome_phylo_infoSEXP) {
@@ -835,6 +793,30 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// make_gamma_mats
+arma::field<arma::mat> make_gamma_mats(const std::vector<uint32>& seq_sizes, const uint32& gamma_size_, const double& shape);
+RcppExport SEXP _jackalope_make_gamma_mats(SEXP seq_sizesSEXP, SEXP gamma_size_SEXP, SEXP shapeSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const std::vector<uint32>& >::type seq_sizes(seq_sizesSEXP);
+    Rcpp::traits::input_parameter< const uint32& >::type gamma_size_(gamma_size_SEXP);
+    Rcpp::traits::input_parameter< const double& >::type shape(shapeSEXP);
+    rcpp_result_gen = Rcpp::wrap(make_gamma_mats(seq_sizes, gamma_size_, shape));
+    return rcpp_result_gen;
+END_RCPP
+}
+// check_gamma_mats
+void check_gamma_mats(const std::vector<arma::mat>& mats, const std::vector<uint32>& seq_sizes);
+RcppExport SEXP _jackalope_check_gamma_mats(SEXP matsSEXP, SEXP seq_sizesSEXP) {
+BEGIN_RCPP
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const std::vector<arma::mat>& >::type mats(matsSEXP);
+    Rcpp::traits::input_parameter< const std::vector<uint32>& >::type seq_sizes(seq_sizesSEXP);
+    check_gamma_mats(mats, seq_sizes);
+    return R_NilValue;
+END_RCPP
+}
 // sub_TN93
 List sub_TN93(std::vector<double> pi_tcag, const double& alpha_1, const double& alpha_2, const double& beta);
 RcppExport SEXP _jackalope_sub_TN93(SEXP pi_tcagSEXP, SEXP alpha_1SEXP, SEXP alpha_2SEXP, SEXP betaSEXP) {
@@ -942,6 +924,24 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// add_ssites_cpp
+SEXP add_ssites_cpp(SEXP& ref_genome_ptr, const std::vector<arma::mat>& seg_sites, const arma::mat& Q, const std::vector<double>& pi_tcag, const std::vector<double>& insertion_rates, const std::vector<double>& deletion_rates, uint32 n_threads, const bool& show_progress);
+RcppExport SEXP _jackalope_add_ssites_cpp(SEXP ref_genome_ptrSEXP, SEXP seg_sitesSEXP, SEXP QSEXP, SEXP pi_tcagSEXP, SEXP insertion_ratesSEXP, SEXP deletion_ratesSEXP, SEXP n_threadsSEXP, SEXP show_progressSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP& >::type ref_genome_ptr(ref_genome_ptrSEXP);
+    Rcpp::traits::input_parameter< const std::vector<arma::mat>& >::type seg_sites(seg_sitesSEXP);
+    Rcpp::traits::input_parameter< const arma::mat& >::type Q(QSEXP);
+    Rcpp::traits::input_parameter< const std::vector<double>& >::type pi_tcag(pi_tcagSEXP);
+    Rcpp::traits::input_parameter< const std::vector<double>& >::type insertion_rates(insertion_ratesSEXP);
+    Rcpp::traits::input_parameter< const std::vector<double>& >::type deletion_rates(deletion_ratesSEXP);
+    Rcpp::traits::input_parameter< uint32 >::type n_threads(n_threadsSEXP);
+    Rcpp::traits::input_parameter< const bool& >::type show_progress(show_progressSEXP);
+    rcpp_result_gen = Rcpp::wrap(add_ssites_cpp(ref_genome_ptr, seg_sites, Q, pi_tcag, insertion_rates, deletion_rates, n_threads, show_progress));
+    return rcpp_result_gen;
+END_RCPP
+}
 
 static const R_CallMethodDef CallEntries[] = {
     {"_jackalope_merge_sequences_cpp", (DL_FUNC) &_jackalope_merge_sequences_cpp, 1},
@@ -964,9 +964,6 @@ static const R_CallMethodDef CallEntries[] = {
     {"_jackalope_write_vcf_cpp", (DL_FUNC) &_jackalope_write_vcf_cpp, 5},
     {"_jackalope_make_mutation_sampler_base", (DL_FUNC) &_jackalope_make_mutation_sampler_base, 4},
     {"_jackalope_make_mutation_sampler_chunk_base", (DL_FUNC) &_jackalope_make_mutation_sampler_chunk_base, 5},
-    {"_jackalope_add_coal_sites_cpp", (DL_FUNC) &_jackalope_add_coal_sites_cpp, 8},
-    {"_jackalope_make_gamma_mats", (DL_FUNC) &_jackalope_make_gamma_mats, 3},
-    {"_jackalope_check_gamma_mats", (DL_FUNC) &_jackalope_check_gamma_mats, 2},
     {"_jackalope_phylo_info_to_trees", (DL_FUNC) &_jackalope_phylo_info_to_trees, 1},
     {"_jackalope_phylo_info_to_trees_chunk", (DL_FUNC) &_jackalope_phylo_info_to_trees_chunk, 1},
     {"_jackalope_evolve_seqs", (DL_FUNC) &_jackalope_evolve_seqs, 6},
@@ -1002,6 +999,8 @@ static const R_CallMethodDef CallEntries[] = {
     {"_jackalope_add_insertion", (DL_FUNC) &_jackalope_add_insertion, 5},
     {"_jackalope_add_deletion", (DL_FUNC) &_jackalope_add_deletion, 5},
     {"_jackalope_test_rate", (DL_FUNC) &_jackalope_test_rate, 7},
+    {"_jackalope_make_gamma_mats", (DL_FUNC) &_jackalope_make_gamma_mats, 3},
+    {"_jackalope_check_gamma_mats", (DL_FUNC) &_jackalope_check_gamma_mats, 2},
     {"_jackalope_sub_TN93", (DL_FUNC) &_jackalope_sub_TN93, 4},
     {"_jackalope_sub_JC69", (DL_FUNC) &_jackalope_sub_JC69, 1},
     {"_jackalope_sub_K80", (DL_FUNC) &_jackalope_sub_K80, 2},
@@ -1011,6 +1010,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_jackalope_sub_GTR", (DL_FUNC) &_jackalope_sub_GTR, 2},
     {"_jackalope_sub_UNREST", (DL_FUNC) &_jackalope_sub_UNREST, 1},
     {"_jackalope_using_openmp", (DL_FUNC) &_jackalope_using_openmp, 0},
+    {"_jackalope_add_ssites_cpp", (DL_FUNC) &_jackalope_add_ssites_cpp, 8},
     {NULL, NULL, 0}
 };
