+++ src/vars_ssites.cpp
@@ -11,53 +11,16 @@
 
 
 #include "jackalope_types.h"
-#include "mevo.h"
+#include "mutator_type.h"
+#include "mutator.h"
 #include "seq_classes_var.h"  // Var* classes
 #include "pcg.h"  // pcg seeding
 #include "alias_sampler.h"  // alias method of sampling
 #include "weighted_reservoir.h"  // weighted reservoir sampling
-#include "mevo_gammas.h"  // SequenceGammas class
+#include "site_var.h"  // SequenceGammas class
 #include "util.h"  // thread_check
 
 using namespace Rcpp;
-
-
-
-
-
-// Wrapper to make non-chunked version available from R
-
-//[[Rcpp::export]]
-SEXP make_mutation_sampler_base(const arma::mat& Q,
-                                const std::vector<double>& pi_tcag,
-                                const std::vector<double>& insertion_rates,
-                                const std::vector<double>& deletion_rates) {
-
-    XPtr<MutationSampler> out =
-        make_mutation_sampler_base_<MutationSampler,LocationSampler>(
-                Q, pi_tcag, insertion_rates, deletion_rates);
-
-    return out;
-}
-
-// Same thing, but with chunks
-
-//[[Rcpp::export]]
-SEXP make_mutation_sampler_chunk_base(const arma::mat& Q,
-                                      const std::vector<double>& pi_tcag,
-                                      const std::vector<double>& insertion_rates,
-                                      const std::vector<double>& deletion_rates,
-                                      const uint32& chunk_size) {
-
-    XPtr<ChunkMutationSampler> out =
-        make_mutation_sampler_base_<ChunkMutationSampler,ChunkLocationSampler>(
-                Q, pi_tcag, insertion_rates, deletion_rates);
-
-    out->location.change_chunk(chunk_size);
-
-    return out;
-}
-
 
 
 
@@ -76,9 +39,9 @@ MutationTypeSampler make_type_sampler(const arma::mat& Q,
     std::vector<double> q_tcag;
 
     /*
-    (1) Combine substitution, insertion, and deletion rates into a single vector
-    (2) Fill the `q_tcag` vector with mutation rates for each nucleotide
-    */
+     (1) Combine substitution, insertion, and deletion rates into a single vector
+     (2) Fill the `q_tcag` vector with mutation rates for each nucleotide
+     */
     fill_probs_q_tcag(probs, q_tcag, Q, pi_tcag, insertion_rates, deletion_rates);
 
     // Now filling in mut_lengths vector
@@ -93,17 +56,18 @@ MutationTypeSampler make_type_sampler(const arma::mat& Q,
 
 
 
+
 //' Add mutations at segregating sites for one sequence from coalescent simulation output.
 //'
 //' @noRd
 //'
-void add_one_seq_sites(VarSet& var_set,
-                       const RefGenome& ref_genome,
-                       const uint32& seq_i,
-                       const arma::mat& ss_i,
-                       MutationTypeSampler& type_sampler,
-                       AliasStringSampler<std::string>& insert_sampler,
-                       pcg64& eng) {
+void add_one_seq_ssites(VarSet& var_set,
+                        const RefGenome& ref_genome,
+                        const uint32& seq_i,
+                        const arma::mat& ss_i,
+                        MutationTypeSampler& type_sampler,
+                        AliasStringSampler<std::string>& insert_sampler,
+                        pcg64& eng) {
 
     uint32 pos;
     std::string nts; // <-- for insertions
@@ -156,14 +120,14 @@ void add_one_seq_sites(VarSet& var_set,
 //' @noRd
 //'
 //[[Rcpp::export]]
-SEXP add_coal_sites_cpp(SEXP& ref_genome_ptr,
-                        const std::vector<arma::mat>& seg_sites,
-                        const arma::mat& Q,
-                        const std::vector<double>& pi_tcag,
-                        const std::vector<double>& insertion_rates,
-                        const std::vector<double>& deletion_rates,
-                        uint32 n_threads,
-                        const bool& show_progress) {
+SEXP add_ssites_cpp(SEXP& ref_genome_ptr,
+                    const std::vector<arma::mat>& seg_sites,
+                    const arma::mat& Q,
+                    const std::vector<double>& pi_tcag,
+                    const std::vector<double>& insertion_rates,
+                    const std::vector<double>& deletion_rates,
+                    uint32 n_threads,
+                    const bool& show_progress) {
 
     XPtr<RefGenome> ref_genome(ref_genome_ptr);
 
@@ -192,7 +156,7 @@ SEXP add_coal_sites_cpp(SEXP& ref_genome_ptr,
     // Type and insertion samplers:
     MutationTypeSampler type = make_type_sampler(Q, pi_tcag, insertion_rates,
                                                  deletion_rates);
-    AliasStringSampler<std::string> insert(mevo::bases, pi_tcag);
+    AliasStringSampler<std::string> insert("TCAG", pi_tcag);
 
     std::vector<uint64> active_seeds;
 
@@ -216,7 +180,7 @@ SEXP add_coal_sites_cpp(SEXP& ref_genome_ptr,
         if (prog_bar.is_aborted() || prog_bar.check_abort()) status_code = -1;
         if (status_code != 0) continue;
 
-        add_one_seq_sites(*var_set, *ref_genome, i, seg_sites[i], type, insert, eng);
+        add_one_seq_ssites(*var_set, *ref_genome, i, seg_sites[i], type, insert, eng);
 
         prog_bar.increment((*ref_genome)[i].size());
